System.PosixCompat

Description

The unix-compat package provides portable implementations of parts of the unix package. On POSIX system it re-exports operations from the unix package, on other platforms it emulates the operations as far as possible.

Documentation

module System.PosixCompat.Files

module System.PosixCompat.Process

module System.PosixCompat.Temp

module System.PosixCompat.Time

module System.PosixCompat.Types

module System.PosixCompat.Unistd

usingPortableImpl :: Bool

True if unix-compat is using its portable implementation, or False if the unix package is simply being re-exported.