System.PosixCompat.Temp

Description

This module makes the operations exported by System.Posix.Temp available on all platforms. On POSIX systems it re-exports operations from System.Posix.Temp, on other platforms it emulates the operations as far as possible.

Documentation

mkstemp :: String -> IO (FilePath, Handle)

mkstemp - make a unique filename and open it for reading/writing (only safe on GHC & Hugs). The returned FilePath is the (possibly relative) path of the created file, which is padded with 6 random characters.