System.PosixCompat.Extensions

Description

This module provides some functions not present in the unix package.

Device IDs.

type CMajor = CUInt

type CMinor = CUInt

deviceMajor :: DeviceID -> CMajor

Gets the major number from a DeviceID for a device file.

The portable implementation always returns 0.

deviceMinor :: DeviceID -> CMinor

Gets the minor number from a DeviceID for a device file.

The portable implementation always returns 0.

makeDeviceID :: CMajor -> CMinor -> DeviceID

Creates a DeviceID for a device file given a major and minor number.