System.PosixCompat.Time

Description

This module makes the operations exported by System.Posix.Time available on all platforms. On POSIX systems it re-exports operations from System.Posix.Time, on other platforms it emulates the operations as far as possible.

Documentation

epochTime :: IO EpochTime

epochTime calls time to obtain the number of seconds that have elapsed since the epoch (Jan 01 00:00:00 GMT 1970).